Improve examples and introduce reconnect example
The reconnect example is a test for doing reconnect problems when partners shutdown their devices
