+++ docker-compose.yaml
@@ -11,14 +11,13 @@ services:
       VPN_LOCAL_NETWORK: 10.101.0.1
       VPN_REMOTE_PEER: strongswan2
       VPN_REMOTE_NETWORK: 10.102.0.1
-      STRONG_DUCKLING_ARGS: "--listen :8000 --tcp-checker nodejs:10.101.0.1:8090 --vici-socket /var/run/charon.vici"
     volumes:
       - type: bind
         source: ./strong-duckling-linux
         target: /strong-duckling
-    ports:
-      - 9001:8000
-      - 9002:8080
+      - type: bind
+        source: ./examples/${EXAMPLE}/left/
+        target: /config
   strongswan2:
     build:
       context: strongswan
@@ -30,14 +29,13 @@ services:
       VPN_LOCAL_NETWORK: 10.102.0.1
       VPN_REMOTE_PEER: strongswan1
       VPN_REMOTE_NETWORK: 10.101.0.1
-      STRONG_DUCKLING_ARGS: "--listen :8000 --tcp-checker 10.101.0.1:8090 --vici-socket /var/run/charon.vici"
-    ports:
-      - 10001:8000
-      - 10002:8080
     volumes:
       - type: bind
         source: ./strong-duckling-linux
         target: /strong-duckling
+      - type: bind
+        source: ./examples/${EXAMPLE}/right
+        target: /config
 
   prometheus:
     image: prom/prometheus:v2.15.2
