+++ examples/simple/left/supervisord.conf
@@ -1,0 +1,39 @@
+[supervisord]
+nodaemon=true
+
+[unix_http_server]
+file=/var/run/supervisor.sock
+chmod=0700
+
+[rpcinterface:supervisor]
+supervisor.rpcinterface_factory = supervisor.rpcinterface:make_main_rpcinterface
+
+[supervisorctl]
+serverurl=unix:///var/run/supervisor.sock
+
+[program:charon]
+command=/prefix-log /usr/sbin/charon-systemd
+stdout_logfile=/dev/stdout
+stdout_logfile_maxbytes=0
+stderr_logfile=/dev/stderr
+stderr_logfile_maxbytes=0
+
+[program:http-server]
+command=/prefix-log node /server.js
+stdout_logfile=/dev/stdout
+stdout_logfile_maxbytes=0
+stderr_logfile=/dev/stderr
+stderr_logfile_maxbytes=0
+
+[program:strong-duckling]
+command=/prefix-log /strong-duckling.sh --listen :8000 --tcp-checker {{ .Env.VPN_REMOTE_NETWORK }}:8080 --vici-socket /var/run/charon.vici || exit 1
+stdout_logfile=/dev/stdout
+stdout_logfile_maxbytes=0
+stderr_logfile=/dev/stderr
+stderr_logfile_maxbytes=0
+
+[program:initialize-script]
+command=bash -c "sleep 2; /usr/sbin/swanctl --load-all --noprompt"
+startsecs=0
+autorestart=false
+startretries=1
