+++ examples/simple/left/swanctl.conf
@@ -1,0 +1,43 @@
+connections {
+  gw-gw {
+    local_addrs  = {{ .Env.IP }},{{ .Env.VPN_LOCAL_PEER }}
+    remote_addrs = {{ .Env.VPN_REMOTE_PEER }}
+
+    local {
+      auth = psk
+      id = {{ .Env.VPN_LOCAL_PEER }}
+    }
+    remote {
+      auth = psk
+      id = {{ .Env.VPN_REMOTE_PEER }}
+    }
+    children {
+      net-net-0 {
+        local_ts = {{ .Env.VPN_LOCAL_NETWORK }}/32
+        remote_ts = {{ .Env.VPN_REMOTE_NETWORK }}/32
+        updown = /usr/libexec/ipsec/_updown iptables
+
+        rekey_time = 5400
+        rekey_bytes = 500000000
+        rekey_packets = 1000000
+        esp_proposals = aes256-sha256-ecp384
+
+        start_action = start
+        close_action = start
+        dpd_action = start
+      }
+    }
+    version = 2
+    mobike = no
+    reauth_time = 10800
+    proposals = aes256-sha256-ecp384
+  }
+}
+
+secrets {
+  ike-1 {
+    id-local = {{ .Env.VPN_LOCAL_PEER }}
+    id-remote = {{ .Env.VPN_REMOTE_PEER }}
+    secret = "123456"
+  }
+}
