+++ strongswan/strong-duckling.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+nodemon --signal SIGTERM --watch /strong-duckling -x "/strong-duckling $@ || exit 1"
